requirements.txt
aiohttp~=3.8.1
aiobungie~=0.2.6
# ssl
urllib3
motor
dnspython
pymongo~=3.12.3
discord.py~=2.0.1
pytz